Implement Analytics tracking code. This was previously handled with the AdServer.php. Since it doesn't change per user or per wiki, there is no reason for this, so let's save a call and implement it statically.
This original code called AdServer like this:
AdServer::getInstance()->getAd('js_bot1');

Which produced this javascript:

<script src="http://www.google-analytics.com/urchin.js" type="text/javascript"></script>
<script type="text/javascript"> _uff=0;_uacct="UA-288915-1";_userv=1;urchinTracker();_userv=1;</script>
<script type="text/javascript" src="http://edge.quantserve.com/quant.js"></script>
<script type="text/javascript">_qacct="p-8bG6eLqkH6Avk";quantserve();</script>

Note that these script calls blocked the page, and would not allow it to render
if either of these two servers were down or slow.

The new code downloads them asynchronously and then fires of the requests for the gifs
when the code is downloaded, allowing the user to see content while the analytics code
is downloading.